refactor: Restructure and organize API files for better separation
