+++ src/components/modal/task-detail/TaskDetail.tsx
@@ -1,6 +1,6 @@
 import { useEffect, useState } from "react";
 import { TaskCardDetail } from "@/lib/types";
-import { fetchTaskCardDetail } from "@/lib/apis/taskcard";
+import { fetchTaskCardDetail } from "@/lib/apis/cards";
 import { TOKEN_1 } from "@/lib/constants/tokens";
 import Modal from "@/components/common/modal/Modal";
