+++ src/lib/apis/columns.ts
@@ -1,0 +1,19 @@
+import { BASE_URL } from "@/lib/constants/urls";
+
+export async function fetchColumnList({
+  token,
+  id,
+}: {
+  token: string;
+  id: string;
+}) {
+  const res = await fetch(`${BASE_URL}/columns?dashboardId=${id}`, {
+    headers: {
+      Accept: "application/json",
+      Authorization: `Bearer ${token}`,
+    },
+    cache: "no-store",
+  });
+
+  return res.json();
+}
